requirements.txt
aiohttp
aiofiles
aioshutil
anytree
apscheduler
aria2p
asyncio
beautifulsoup4
bencoding
cloudscraper
dnspython
feedparser
flask
gevent
gunicorn
git+https://github.com/zevtyardt/lk21.git
httpx
lxml
motor
mutagen
natsort
pillow
psutil
pybase64
pymongo
pyrogram
python-dotenv
python-magic
requests
telegraph
tenacity
tgcrypto
uvloop
xattr
yt-dlp